+++ pysgrs/__init__.py
@@ -5,7 +5,7 @@ from pysgrs.alphabets import *
 from pysgrs.ciphers import *
 from pysgrs.toolbox import *
 
-__version__ = "0.0.21"
+__version__ = "0.0.22"
 
 warnings.simplefilter("always")