+++ pysgrs/interfaces/__init__.py
@@ -1,2 +1,3 @@
 from pysgrs.interfaces.alphabet import *
 from pysgrs.interfaces.cipher import *
+from pysgrs.interfaces.factory import *